creating all consumers
